# docs/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Music Gen AI API
  version: 1.0.0
  description: |
    Production-ready AI music generation API that creates high-quality music from text prompts using transformer architecture with EnCodec audio tokenization.

    ## Features
    - **Text-to-Music Generation**: Generate music from natural language descriptions
    - **Multiple Model Support**: Various model sizes for different use cases
    - **Real-time Streaming**: Stream music generation progress
    - **High-Quality Audio**: 32kHz sampling rate with EnCodec compression
    - **Rate Limiting**: Tier-based usage controls
    - **Authentication**: JWT-based secure access
    - **Monitoring**: Comprehensive metrics and health endpoints

    ## Getting Started
    1. Sign up for an API key at [https://musicgen.ai/signup](https://musicgen.ai/signup)
    2. Authenticate using JWT tokens or API keys
    3. Start generating music with simple text prompts
    4. Monitor your usage and billing in the dashboard

    ## Support
    - **Documentation**: [https://docs.musicgen.ai](https://docs.musicgen.ai)
    - **Support**: [support@musicgen.ai](mailto:support@musicgen.ai)
    - **Status Page**: [https://status.musicgen.ai](https://status.musicgen.ai)

  termsOfService: https://musicgen.ai/terms
  contact:
    name: Music Gen AI Support
    url: https://musicgen.ai/support
    email: support@musicgen.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - ApiKeyAuth: []
  - JwtAuth: []
paths:
...
  /generate:
    post:
      tags:
        - Music Generation
      summary: Generate music from text
      description: |
        Generate music from a text prompt using AI models.
        
        **Rate Limits by Tier**:
        - Free: 10 generations/hour, max 30s duration
        - Pro: 100 generations/hour, max 5min duration
        - Enterprise: Unlimited, max 10min duration
        
        **Generation Time**: Typically 2-3x the audio duration
        
        **Supported Models**:
        - `musicgen-small`: Fast, good quality (default)
        - `musicgen-medium`: Balanced speed/quality  
        - `musicgen-large`: Best quality, slower
        - `musicgen-melody`: Melody conditioning support
      operationId: generateMusic
      security:
        - ApiKeyAuth: []
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationRequest'
            examples:
              simple_prompt:
                summary: Simple text prompt
                value:
                  prompt: "Upbeat jazz with saxophone solo"
                  duration: 30
                  model: "musicgen-small"
              advanced_prompt:
                summary: Advanced configuration
                value:
                  prompt: "Ambient electronic music with soft synth pads and gentle arpeggios"
                  duration: 60
                  model: "musicgen-medium"
                  temperature: 0.8
                  top_k: 250
                  top_p: 0.9
                  seed: 42
                  guidance_scale: 7.5
                  format: "wav"
                  sample_rate: 32000
              melody_conditioning:
                summary: Melody conditioning
                value:
                  prompt: "Classical piano piece in the style of Chopin"
                  duration: 120
                  model: "musicgen-melody"
                  melody_url: "https://example.com/melody.wav"
                  melody_conditioning_scale: 0.7
      responses:
        '202':
          description: Generation request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationResponse'
              example:
                task_id: task_abc123def456
                status: queued
                estimated_completion_time: "2024-01-15T10:32:00Z"
                queue_position: 3
                message: "Your music generation request has been queued"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
  /generate/{task_id}:
    get:
      tags:
        - Music Generation
      summary: Get generation status
      description: |
        Get the current status of a music generation task.
        
        **Status Values**:
        - `queued`: Task is waiting in queue
        - `processing`: AI model is generating music
        - `completed`: Generation finished successfully
        - `failed`: Generation failed with error
        - `cancelled`: Task was cancelled by user
      operationId: getGenerationStatus
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^task_[a-zA-Z0-9]{12}$'
          example: task_abc123def456
          description: Unique task identifier
      responses:
        '200':
          description: Task status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
              examples:
                queued:
                  summary: Task in queue
                  value:
                    task_id: task_abc123def456
                    status: queued
                    progress: 0
                    queue_position: 3
                    estimated_completion_time: "2024-01-15T10:32:00Z"
                    created_at: "2024-01-15T10:30:00Z"
                processing:
                  summary: Task processing
                  value:
                    task_id: task_abc123def456
                    status: processing
                    progress: 65
                    estimated_completion_time: "2024-01-15T10:31:30Z"
                    created_at: "2024-01-15T10:30:00Z"
                    started_at: "2024-01-15T10:30:15Z"
                completed:
                  summary: Task completed
                  value:
                    task_id: task_abc123def456
                    status: completed
                    progress: 100
                    audio_url: "https://cdn.musicgen.ai/audio/task_abc123def456.wav"
                    duration: 30.5
                    file_size: 1572864
                    metadata:
                      model: musicgen-small
                      prompt: "Upbeat jazz with saxophone solo"
                      generation_time: 45.2
                      seed: 42
                    created_at: "2024-01-15T10:30:00Z"
                    started_at: "2024-01-15T10:30:15Z"
                    completed_at: "2024-01-15T10:31:00Z"
                failed:
                  summary: Task failed
                  value:
                    task_id: task_abc123def456
                    status: failed
                    progress: 30
                    error:
                      code: GENERATION_FAILED
                      message: "Model failed to generate audio"
                      details: "GPU out of memory error"
                    created_at: "2024-01-15T10:30:00Z"
                    started_at: "2024-01-15T10:30:15Z"
                    failed_at: "2024-01-15T10:30:45Z"
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Include your API key in the `X-API-Key` header.
        
        Example: `X-API-Key: sk_live_1234567890abcdef`
        
        Get your API key from the [dashboard](https://musicgen.ai/dashboard).

    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. Include the token in the `Authorization` header.
        
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
        
        Obtain tokens via the `/auth/login` endpoint.
  schemas:
...
    GenerationRequest:
      type: object
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 500
          description: Text description of the music to generate
          example: "Upbeat jazz with saxophone solo"
        duration:
          type: number
          minimum: 1
          maximum: 600
          description: Duration in seconds (max depends on user tier)
          example: 30
        model:
          type: string
          enum: [musicgen-small, musicgen-medium, musicgen-large, musicgen-melody]
          default: musicgen-small
          description: AI model to use for generation
        temperature:
          type: number
          minimum: 0.1
          maximum: 2.0
          default: 1.0
          description: Sampling temperature (higher = more creative)
        top_k:
          type: integer
          minimum: 1
          maximum: 1000
          default: 250
          description: Top-k sampling parameter
        top_p:
          type: number
          minimum: 0.1
          maximum: 1.0
          default: 0.9
          description: Top-p sampling parameter
        seed:
          type: integer
          minimum: 0
          maximum: 2147483647
          description: Random seed for reproducible generation
        guidance_scale:
          type: number
          minimum: 1.0
          maximum: 20.0
          default: 7.5
          description: How closely to follow the prompt
        format:
          type: string
          enum: [wav, mp3, flac]
          default: wav
          description: Output audio format
        sample_rate:
          type: integer
          enum: [16000, 22050, 24000, 32000, 44100, 48000]
          default: 32000
          description: Output sample rate in Hz
        melody_url:
          type: string
          format: uri
          description: URL to melody file for conditioning (musicgen-melody only)
        melody_conditioning_scale:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 0.5
          description: Strength of melody conditioning
      required: [prompt]
    GenerationResponse:
      type: object
      properties:
        task_id:
          type: string
          description: Unique task identifier
          pattern: '^task_[a-zA-Z0-9]{12}$'
          example: task_abc123def456
        status:
          type: string
          enum: [queued]
          description: Initial task status
        estimated_completion_time:
          type: string
          format: date-time
          description: Estimated completion time
        queue_position:
          type: integer
          description: Position in generation queue
        message:
          type: string
          description: Human-readable status message
      required: [task_id, status]
    TaskStatus:
      type: object
      properties:
        task_id:
          type: string
          description: Unique task identifier
        status:
          type: string
          enum: [queued, processing, completed, failed, cancelled]
          description: Current task status
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Completion percentage
        queue_position:
          type: integer
          description: Position in queue (only for queued tasks)
        estimated_completion_time:
          type: string
          format: date-time
          description: Estimated completion time
        audio_url:
          type: string
          format: uri
          description: URL to generated audio file (only for completed tasks)
        duration:
          type: number
          description: Audio duration in seconds
        file_size:
          type: integer
          description: File size in bytes
        metadata:
          type: object
          description: Generation metadata
          properties:
            model:
              type: string
              description: Model used for generation
            prompt:
              type: string
              description: Original text prompt
            generation_time:
              type: number
              description: Time taken to generate in seconds
            seed:
              type: integer
              description: Random seed used
        error:
          $ref: '#/components/schemas/Error'
        created_at:
          type: string
          format: date-time
          description: Task creation time
        started_at:
          type: string
          format: date-time
          description: Task start time
        completed_at:
          type: string
          format: date-time
          description: Task completion time
        failed_at:
          type: string
          format: date-time
          description: Task failure time
      required: [task_id, status, progress, created_at]
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: "The prompt field is required"
        details:
          type: string
          description: Additional error details
        request_id:
          type: string
          description: Unique request identifier for debugging
          example: req_1234567890abcdef
      required: [code, message]
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation_error:
              summary: Validation error
              value:
                code: VALIDATION_ERROR
                message: "The prompt field is required"
                request_id: req_1234567890abcdef
            invalid_duration:
              summary: Invalid duration
              value:
                code: INVALID_DURATION
                message: "Duration must be between 1 and 300 seconds for your tier"
                details: "Free tier is limited to 30 seconds maximum"
                request_id: req_1234567890abcdef
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_auth:
              summary: Missing authentication
              value:
                code: AUTHENTICATION_REQUIRED
                message: "Authentication is required to access this resource"
                request_id: req_1234567890abcdef
            invalid_token:
              summary: Invalid token
              value:
                code: INVALID_TOKEN
                message: "The provided authentication token is invalid or expired"
                request_id: req_1234567890abcdef
            expired_token:
              summary: Expired token
              value:
                code: TOKEN_EXPIRED
                message: "Authentication token has expired. Please refresh your token."
                request_id: req_1234567890abcdef
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            task_not_found:
              summary: Task not found
              value:
                code: TASK_NOT_FOUND
                message: "The specified task could not be found"
                request_id: req_1234567890abcdef
            model_not_found:
              summary: Model not found
              value:
                code: MODEL_NOT_FOUND
                message: "The specified model is not available"
                request_id: req_1234567890abcdef
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            rate_limit_exceeded:
              summary: Rate limit exceeded
              value:
                code: RATE_LIMIT_EXCEEDED
                message: "Rate limit exceeded. Please wait before making another request."
                details: "You have exceeded the rate limit of 10 requests per minute"
                request_id: req_1234567890abcdef
    ServiceUnavailableError:
      description: Service temporarily unavailable
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            maintenance:
              summary: Maintenance mode
              value:
                code: MAINTENANCE_MODE
                message: "Service is temporarily unavailable due to maintenance"
                details: "Maintenance window: 2024-01-15 02:00-04:00 UTC"
                request_id: req_1234567890abcdef
            overloaded:
              summary: Service overloaded
              value:
                code: SERVICE_OVERLOADED
                message: "Service is temporarily overloaded. Please try again later."
                request_id: req_1234567890abcdef
...